Update forms.py
checking if the field is an empty string and the length if the sp_params is only 1. IE no params for the stored procedure. Skips the key error and call with none, which works. I guess another solution would be to wrap the sp get params sql call could return none and the for loop would never run but the params(empty array) would be returned.